feat: Add TodosService to TodosModule
This commit adds the TodosService to the TodosModule in the todos.module.ts file. By including the TodosService as a provider, it enables the module to utilize the functionality provided by the service. This enhancement improves the functionality of the application by introducing a new service and making it available within the module.