feat: Integrate TodosService into todos resolver
This commit updates the todos resolver in the todos.resolver.ts file to utilize the TodosService. Instead of returning an empty array of todos, the resolver now calls the findAll() method of the TodosService to retrieve the list of todos. This enhancement improves the functionality of the application by integrating the TodosService into the resolver and ensuring that the actual data is returned.