feat: Add findById query to todos resolver
This commit adds a new query called `findById` to the todos resolver in the todos.resolver.ts file. The query allows the user to retrieve a single todo by providing the todo's ID as an argument. This enhancement improves the functionality of the application by providing a way to fetch a specific todo from the list of todos.