feat: Add createTodo method to TodosService
This commit adds a new method called `createTodo` to the `TodosService` in the `todos.service.ts` file. The method takes a `CreateTodoInput` object as input and creates a new todo with the provided data. It generates a unique ID for the todo and sets the `done` property to `false`. This enhancement improves the functionality of the application by allowing the creation of new todos.