feat: Add deleteTodo method to TodosService
This commit adds a new method called deleteTodo to the TodosService in the todos.service.ts file. The deleteTodo method takes an id as input and removes the corresponding todo from the todos array. It also includes error handling to throw a NotFoundException if the todo with the given id is not found. This enhancement improves the functionality of the application by allowing the deletion of todos.