feat: Add statusArgs parameter to findAll method in TodosResolver
This commit adds a new parameter called statusArgs of type StatusArgs to the findAll method in the TodosResolver. The statusArgs parameter allows for filtering the list of todos based on their status. This enhancement improves the functionality of the application by providing the ability to retrieve todos with specific statuses.