feat: Add status filtering to todos query
This commit modifies the todos query in the schema.gql file to include a new optional parameter called status. The status parameter allows for filtering the list of todos based on their status. This enhancement improves the functionality of the application by providing the ability to retrieve todos with specific statuses.