feat: Add AgregationType for aggregating todo data
This commit adds a new file, aggregations.type.ts, which defines the AgregationType class. The AgregationType class is an ObjectType that represents aggregated data for todos. It includes fields for the total number of todos, the number of pending todos, the number of completed todos, and a deprecated field for the total number of completed todos. This addition enhances the application by providing a structured way to retrieve and display aggregated todo data.